+++ lib/catalyst_builder.dart
@@ -3,6 +3,8 @@ library catalyst_builder;
 export 'src/annotation/annotation.dart';
 export 'src/exception/exception.dart';
 export 'src/service_descriptor.dart';
+export 'src/enhanceable_provider.dart';
+export 'src/lazy_service_descriptor.dart';
 export 'src/service_lifetime.dart';
 export 'src/service_provider.dart';
 export 'src/service_registry.dart';